search bar works for all and covers task and event
