+++ lib/common/database.dart
@@ -584,7 +584,7 @@ class DatabaseService {
   /// Get List of all Tasks with the given tag
   /// Returns a List of Tasks
   /// Returns empty List if tag doesn't exist
-  Future<List<Task>> getTasksWithTag(String tagName, {int limit = 100}) async {
+  Future<List<Task>> getTasksWithTag(String tagName, int limit) async {
     List<Task> out = [];
     Tag tag;
 
@@ -604,8 +604,7 @@ class DatabaseService {
   /// Get ID of all Events with the given tag
   /// Returns a list of IDs
   /// Returns empty list if tag doesn't exist
-  Future<List<Event>> getEventsWithTag(String tagName,
-      {int limit = 100}) async {
+  Future<List<Event>> getEventsWithTag(String tagName, int limit) async {
     List<Event> out = [];
     Tag tag;
