+++ lib/common/view/topbar.dart
@@ -1,5 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:planner/common/database.dart';
+import 'package:planner/common/time_management.dart';
 import 'package:planner/models/event.dart';
+import 'package:planner/models/task.dart';
+import 'package:planner/models/undertaking.dart';
 import 'package:planner/view/dayView.dart';
 import 'package:planner/view/monthView.dart';
 import 'package:planner/view/monthlyTaskView.dart';
@@ -13,6 +17,101 @@ RoundedRectangleBorder roundedRectangleBackground =
   bottomLeft: Radius.circular(20),
   bottomRight: Radius.circular(20),
 ));
+
+///A void function that shows a dialog with a search bar to search for tasks.
+void showSearchBar(BuildContext context) {
+  TextEditingController searchController = TextEditingController();
+  DatabaseService db = DatabaseService();
+  
+  showDialog(
+    context: context,
+    builder: (context) {
+      return AlertDialog(
+        title: const Text('Search'),
+        content: TextField(
+          controller: searchController,
+          decoration: const InputDecoration(),
+        ),
+        actions: <Widget>[
+          TextButton(
+            child: const Text('Cancel'),
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+          ),
+          TextButton(
+            child: const Text('Search'),
+            onPressed: () async {
+              String searchQuery = searchController.text;
+              List<Task> searchTask = await db.searchAllTask(searchQuery);
+              List<Event> searchEvent = await db.searchAllEvent(searchQuery);
+              showTaskDetailsDialog(searchQuery, searchTask, searchEvent, context);
+            },
+          ),
+        ],
+      );
+    },
+  );
+}
+
+///A void function that searches in a query and a list of tasks to query from
+///Returns a list of tasks with informations of each tasks
+void showTaskDetailsDialog(String searchQuery, List<Task> tasks, List<Event> events, context) {
+  showDialog(
+    context: context,
+    builder: (BuildContext context) {
+      return AlertDialog(
+        title: Text('Results for "$searchQuery"'),
+        content: SingleChildScrollView(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: tasks.map((task) {
+              return Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    '${task.completed ? "✅" : "❌"} ${task.name}',
+                    style: const TextStyle(fontWeight: FontWeight.bold),
+                  ),
+                  Text('  ${task.description}'),
+                  Text(
+                      '  Currently on: ${getDateAsString(task.timeCurrent)}'),
+                  Text(
+                      '  Date created: ${getDateAsString(task.timeCreated)}'),
+                  const Divider(),
+                ],
+              );
+            }).toList() + events.map((event) {
+              return Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    '🕒 ${event.name}',
+                    style: const TextStyle(fontWeight: FontWeight.bold),
+                  ),
+                  Text('  ${event.description}'),
+                  Text(
+                      '  Starts at: ${getTimeAsString(event.timeStart)}'),
+                  Text(
+                      '  Date created: ${getDateAsString(event.timeCreated)}'),
+                  const Divider(),
+                ],
+              );
+            }).toList(),
+          ),
+        ),
+        actions: [
+          TextButton(
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+            child: const Text('Close'),
+          ),
+        ],
+      );
+    },
+  );
+}
 
 AppBar _getTopBarDaily(bool forEvents, BuildContext context, state) {
   return AppBar(
@@ -81,7 +180,7 @@ AppBar _getTopBarDaily(bool forEvents, BuildContext context, state) {
           color: Colors.black,
         ),
         onPressed: () {
-          state.showSearchBar(context);
+          showSearchBar(context);
         },
       ),
     ],
@@ -171,7 +270,7 @@ AppBar _getTopBarWeekly(bool forEvents, BuildContext context, state) {
       IconButton(
         icon: const Icon(Icons.search, color: Colors.black),
         onPressed: () {
-          //showSearchBar();
+          showSearchBar(context);
         },
       ),
       IconButton(
@@ -245,7 +344,7 @@ _getTopBarMonthly(bool forEvents, BuildContext context, state) {
           color: Colors.black,
         ),
         onPressed: () {
-          //showSearchBar(context);
+          showSearchBar(context);
         },
       ),
     ],
