+++ lib/view/taskView.dart
@@ -177,83 +177,6 @@ class TaskViewState extends State<TaskView> {
     return completer.future;
   }
 
-  ///A void function that shows a dialog with a search bar to search for tasks.
-  void showSearchBar(BuildContext context) {
-    TextEditingController searchController = TextEditingController();
-
-    showDialog(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          title: const Text('Search Tasks'),
-          content: TextField(
-            controller: searchController,
-            decoration: const InputDecoration(),
-          ),
-          actions: <Widget>[
-            TextButton(
-              child: const Text('Cancel'),
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-            ),
-            TextButton(
-              child: const Text('Search'),
-              onPressed: () async {
-                String searchQuery = searchController.text;
-                List<Task> searchTask = await db.searchAllTask(searchQuery);
-                showTaskDetailsDialog(searchQuery, searchTask);
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  ///A void function that searches in a query and a list of tasks to query from
-  ///Returns a list of tasks with informations of each tasks
-  void showTaskDetailsDialog(String searchQuery, List<Task> tasks) {
-    showDialog(
-      context: scaffoldKey.currentState!.context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text('Results for "$searchQuery"'),
-          content: SingleChildScrollView(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: tasks.map((task) {
-                return Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      '${task.completed ? "✅" : "❌"} ${task.name}',
-                      style: const TextStyle(fontWeight: FontWeight.bold),
-                    ),
-                    Text('  ${task.description}'),
-                    Text(
-                        '  Currently on: ${getDateAsString(task.timeCurrent)}'),
-                    Text(
-                        '  Date created: ${getDateAsString(task.timeCreated)}'),
-                    const Divider(),
-                  ],
-                );
-              }).toList(),
-            ),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-              child: const Text('Close'),
-            ),
-          ],
-        );
-      },
-    );
-  }
-
   var scaffoldKey = GlobalKey<ScaffoldState>();
 
   @override
